add:eventのAPI定義書

+++ src/database/migrations/2023_11_28_134713_create_events_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('events', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('community_id');
+            $table->integer('community_id');
             $table->float('gen');
             $table->string('name');
             $table->string('kind');
